refactor(harness): move MeasurementHarness implementation out of package init
Move the MeasurementHarness god class and every module-level helper it
resolves as a bare global (_cuda_sync, PowerThermalSampler,
measure_baseline_power, collect_environment_snapshot[_async],
select_energy_sampler, load_prompts, thermal_floor_wait,
warmup_until_converged, write_timeseries_parquet,
collect_measurement_warnings, estimate_flops_palm[_from_config],
_capture_observed_params_into_output, _check_persistence_mode) into a new
harness/measurement.py. The package __init__.py becomes a pure re-export
shim so every existing import path keeps working.

Tests that patch harness module globals are repointed to
llenergymeasure.harness.measurement.<name> (patching the shim re-export
does not affect measurement.py globals). Method-object patches of
MeasurementHarness.run resolve through the shim to the same class object
and are unchanged.

Update the import-linter ignore for the config-sidecar persistence import
to the measurement-specific path (the package-level ignore no longer
matches now that the importer is harness.measurement).

Zero behaviour change; byte-faithful move apart from the module docstring.
